requirements.txt
streamlit
pandas
folium
requests
openpyxl
xlrd
branca
pdfplumber
python-docx
reportlab
pytesseract
Pillow
streamlit-sortables
streamlit-pdf-viewer
pymupdf
beautifulsoup4
lxml
docxtopdf
fonttools
ddgs
duckduckgo-search